% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_stl.R
\name{loess}
\alias{loess}
\title{Fit a Loess regression.}
\usage{
loess(y, window, degree = 1, jump = 0)
}
\arguments{
\item{y}{input time series.}

\item{window}{}

\item{degree}{}

\item{jump}{}
}
\description{
Fit a Loess regression.
}
\examples{
q<-rjd3stl::stl(rjd3toolkit::ABS$X0.2.09.10.M, 12)
decomp<-q$decomposition
t<-decomp[,'t']
matplot(cbind(t,loess(t, 121)), type='l')
}
