% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_stl.R
\name{mstl}
\alias{mstl}
\title{Title}
\usage{
mstl(
  y,
  periods,
  multiplicative = TRUE,
  swindows = NULL,
  twindow = 0,
  ninnerloop = 1,
  nouterloop = 15,
  nojump = FALSE,
  weight.threshold = 0.001,
  weight.function = c("BIWEIGHT", "UNIFORM", "TRIANGULAR", "EPANECHNIKOV", "TRICUBE",
    "TRIWEIGHT")
)
}
\arguments{
\item{y}{}

\item{periods}{}

\item{multiplicative}{}

\item{swindows}{}

\item{twindow}{}

\item{ninnerloop}{}

\item{nouterloop}{}

\item{nojump}{}

\item{weight.threshold}{}

\item{weight.function}{}
}
\description{
Title
}
\examples{
q<-rjd3stl::mstl(rjd3toolkit::ABS$X0.2.09.10.M, c(12, 25))
decomp<-q$decomposition
}
