% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_stl.R
\name{stl}
\alias{stl}
\title{Title}
\usage{
stl(
  y,
  period,
  multiplicative = TRUE,
  swindow = 7,
  twindow = 0,
  ninnerloop = 1,
  nouterloop = 15,
  nojump = FALSE,
  weight.threshold = 0.001,
  weight.function = c("BIWEIGHT", "UNIFORM", "TRIANGULAR", "EPANECHNIKOV", "TRICUBE",
    "TRIWEIGHT")
)
}
\arguments{
\item{y}{input time series.}

\item{period}{period, any positive real number.}

\item{multiplicative}{Boolean indicating if the decomposition mode is multiplicative (TRUE).}

\item{swindow}{length of seasonal filter.}

\item{twindow}{length of trend filter.}

\item{ninnerloop}{Number of inner loops}

\item{nouterloop}{Number of outer loops (computation of robust weights)}

\item{nojump}{}

\item{weight.threshold}{Weights threshold (in [0, 0.3])}

\item{weight.function}{weights function}
}
\description{
Perform an STL like (based on Loess) decomposition on any periodicity
}
\examples{
q<-rjd3stl::stl(rjd3toolkit::ABS$X0.2.09.10.M, 12)
decomp<-q$decomposition
}
